clean up get_witnesses_for_bill()
